+++ CHANGELOG.md
@@ -20,7 +20,8 @@ as of 2.0.0.
 - Add language filter to combobox for requested languages (#249)
 
 ### Changed
-
+- Move scraper files to `scraper` subfolder and update workflows
+- Initialize new Vue.js project in `zimui` subfolder
 - Simplify Gutenberg scraping (no more rsync, no more fallback URLs / filenames) (#97)
 - Prefer EPUB 3 to EPUB (#235)
 - Do not force the presence of PDF format for all books (#160)